add annbatch streaming path
