+++ src/cellflow/data/__init__.py
@@ -1,6 +1,6 @@
 from cellflow.data._data import BaseDataMixin, ConditionData, PredictionData, TrainingData, ValidationData
-from cellflow.data._dataloader import PredictionSampler, TrainSampler, ValidationSampler
 from cellflow.data._datamanager import DataManager
+from cellflow.data._legacy import PredictionSampler, TrainSampler, ValidationSampler
 
 __all__ = [
     "DataManager",
