+++ src/cellflow/model/__init__.py
@@ -1,3 +1,4 @@
 from cellflow.model._cellflow import CellFlow
+from cellflow.model._cellflow_annbatch import CellFlowAnnbatch
 
-__all__ = ["CellFlow"]
+__all__ = ["CellFlow", "CellFlowAnnbatch"]
