+++ src/cellflow/training/_trainer.py
@@ -6,7 +6,7 @@ import numpy as np
 from numpy.typing import ArrayLike
 from tqdm import tqdm
 
-from cellflow.data._dataloader import OOCTrainSampler, TrainSampler, ValidationSampler
+from cellflow.data._legacy import OOCTrainSampler, TrainSampler, ValidationSampler
 from cellflow.solvers import _genot, _otfm
 from cellflow.training._callbacks import BaseCallback, CallbackRunner
