+++ tests/data/test_cfsampler.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from cellflow.data._dataloader import OOCTrainSampler, PredictionSampler, TrainSampler
+from cellflow.data._legacy import OOCTrainSampler, PredictionSampler, TrainSampler
 from cellflow.data._datamanager import DataManager
 
 
@@ -86,7 +86,7 @@ class TestOOCTrainSampler:
 class TestValidationSampler:
     @pytest.mark.parametrize("n_conditions_on_log_iteration", [None, 1, 3])
     def test_valid_sampler(self, adata_perturbation, n_conditions_on_log_iteration):
-        from cellflow.data._dataloader import ValidationSampler
+        from cellflow.data._legacy import ValidationSampler
         from cellflow.data._datamanager import DataManager
 
         control_key = "control"
