+++ tests/model/test_annbatch_ooc.py
@@ -60,8 +60,8 @@ def _prepare_model_small(cf):
 
 class TestOutOfCore:
     def test_grouped_collection_chunked_trains(self, tmp_path):
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(
             source=_collection(tmp_path, grouped=True),
             sampler_config=SamplerConfig(batch_size=16, chunk_size=4, preload_nchunks=16),
             **_PREP,
@@ -72,9 +72,9 @@ class TestOutOfCore:
         assert cf.solver is not None
 
     def test_ungrouped_collection_chunked_raises(self, tmp_path):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         with pytest.raises(ValueError, match="grouped|groupby"):
-            cf.prepare_annbatch_data(
+            cf.prepare_loaders(
                 source=_collection(tmp_path, grouped=False),
                 sampler_config=SamplerConfig(batch_size=16, chunk_size=4, preload_nchunks=16),
                 **_PREP,
@@ -82,8 +82,8 @@ class TestOutOfCore:
 
     def test_ungrouped_collection_chunk1_ok(self, tmp_path):
         # chunk_size=1 streams per-row → no grouping needed even for an interleaved collection
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(
             source=_collection(tmp_path, grouped=False),
             sampler_config=SamplerConfig(batch_size=16, chunk_size=1, preload_nchunks=16),
             **_PREP,
@@ -91,8 +91,8 @@ class TestOutOfCore:
         assert cf._dataloader.sample()["tgt_cell_data"].shape == (16, 5)
 
     def test_grouped_collection_split_chunked(self, tmp_path):
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(
             source=_collection(tmp_path, grouped=True),
             sampler_config=SamplerConfig(batch_size=16, chunk_size=4, preload_nchunks=16),
             split_by=["drug"],
@@ -103,8 +103,8 @@ class TestOutOfCore:
 
     def test_in_memory_unsorted_source_is_auto_grouped(self):
         # an in-memory (interleaved) AnnData source is grouped automatically → chunk_size>1 works
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(
             source=_adata(interleaved=True),
             sampler_config=SamplerConfig(batch_size=16, chunk_size=4, preload_nchunks=16),
             **_PREP,
@@ -132,8 +132,8 @@ class TestOutOfCore:
         n_runs = 1 + int((np.diff(codes) != 0).sum())
         assert n_runs > 6, f"expected a fragmented collection (>6 runs for 6 categories), got {n_runs}"
 
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(  # chunk_size=4 <= run length 10 → must be accepted
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(  # chunk_size=4 <= run length 10 → must be accepted
             source=dc, sampler_config=SamplerConfig(batch_size=8, chunk_size=4, preload_nchunks=8), **_PREP
         )
         assert cf._dataloader.sample()["tgt_cell_data"].shape == (8, 5)
