+++ tests/model/test_annbatch_split.py
@@ -44,12 +44,12 @@ class TestSplitAnnbatchData:
     """`split_annbatch_data` operates on an already-built Scheme (injected here)."""
 
     def test_without_scheme_raises(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         with pytest.raises(ValueError, match="No annbatch `Scheme`"):
             cf.split_annbatch_data(split_by=["drug"])
 
     def test_split_with_injected_scheme(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         cf._scheme = _toy_scheme()
         df = cf.split_annbatch_data(split_by=["drug"], random_state=0)
         assert list(df.columns) == ["cell_line", "drug", "split"]
@@ -60,10 +60,10 @@ class TestSplitAnnbatchData:
 
 
 class TestPrepareAnnbatchData:
-    """`prepare_annbatch_data` builds Scheme + condition + loaders from an AnnData source."""
+    """`prepare_loaders` builds Scheme + condition + loaders from an AnnData source."""
 
     def _prepare(self, cf, **kwargs):
-        return cf.prepare_annbatch_data(
+        return cf.prepare_loaders(
             source=_toy_adata(),
             sample_rep="X",
             control_key="control",
@@ -73,16 +73,10 @@ class TestPrepareAnnbatchData:
             **kwargs,
         )
 
-    def test_rejects_adata_mode(self, adata_perturbation):
-        with pytest.warns(FutureWarning):
-            cf = cellflow.model.CellFlow(adata_perturbation)
-        with pytest.raises(ValueError, match="without `adata`"):
-            self._prepare(cf)
-
     def test_requires_sampler_config(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         with pytest.raises(ValueError, match="sampler_config` is required"):
-            cf.prepare_annbatch_data(
+            cf.prepare_loaders(
                 source=_toy_adata(),
                 sample_rep="X",
                 control_key="control",
@@ -90,18 +84,18 @@ class TestPrepareAnnbatchData:
             )
 
     def test_builds_scheme_condition_and_loader_without_split(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         self._prepare(cf)
         assert cf._scheme is not None and cf._scheme.root == "pert"
         assert set(cf._annbatch_loaders) == {"train"}  # no split → single "train"
-        assert cf._dataloader is not None  # DAGLoaderTrainSampler wired for train()
+        assert cf._dataloader is not None  # DAGLoaderAdapter wired for train()
         # condition embeddings assembled (drug is categorical → one-hot), data dim from X
         assert set(cf._condition_data) == {"drug"}
         assert cf._data_dim == 5
         assert cf._split_schemes is None
 
     def test_split_builds_per_split_loaders(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         assert self._prepare(cf, split_by=["drug"], split_random_state=0) is None  # prepare_* returns None
         assert set(cf._split_assignment["split"]) == {"train", "val", "test"}
         assert set(cf._annbatch_loaders) == {"train", "val", "test"}
@@ -109,10 +103,10 @@ class TestPrepareAnnbatchData:
         assert cf._dataloader is not None
 
     def test_per_split_sampler_config(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         train = SamplerConfig(batch_size=8, chunk_size=1, preload_nchunks=8)
         small = SamplerConfig(batch_size=4, chunk_size=1, preload_nchunks=4)
-        cf.prepare_annbatch_data(
+        cf.prepare_loaders(
             source=_toy_adata(),
             sample_rep="X",
             control_key="control",
@@ -125,9 +119,9 @@ class TestPrepareAnnbatchData:
         assert cf._annbatch_sampler_configs["val"].batch_size == 4
 
     def test_per_split_missing_split_raises(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         with pytest.raises(ValueError, match="missing config"):
-            cf.prepare_annbatch_data(
+            cf.prepare_loaders(
                 source=_toy_adata(),
                 sample_rep="X",
                 control_key="control",
@@ -138,7 +132,7 @@ class TestPrepareAnnbatchData:
             )
 
     def test_streamed_batch_shapes(self):
-        cf = cellflow.model.CellFlow()
+        cf = cellflow.model.CellFlowAnnbatch()
         self._prepare(cf)
         batch = cf._dataloader.sample()
         assert batch["src_cell_data"].shape == (8, 5)
@@ -151,8 +145,8 @@ class TestPrepareAnnbatchData:
 
         adata = _toy_adata()
         adata.X = sp.csr_matrix(adata.X)
-        cf = cellflow.model.CellFlow()
-        cf.prepare_annbatch_data(
+        cf = cellflow.model.CellFlowAnnbatch()
+        cf.prepare_loaders(
             source=adata,
             sample_rep="X",
             control_key="control",
